// packages/omnivore-polish/src/types/api.ts
export interface NewsletterSubscription {
  id: string
  userId: string
  senderEmail: string
  emailAlias: string
  newsletterEmail?: string | null
  title?: string | null
  description?: string | null
  siteUrl?: string | null
  siteIcon?: string | null
  lastReceivedAt?: string | null
  itemCount: number
  unreadCount?: number | null
  active: boolean
  folder?: string | null
  autoAddLabels?: string[] | null
  unsubscribeMailTo?: string | null
  unsubscribeHttpUrl?: string | null
  createdAt: string
  updatedAt: string
}